requirements.txt
# 深層学習フレームワーク
torch>=2.0.0
torchvision>=0.15.0
pytorch-lightning>=2.0.0

# MLflow
mlflow>=2.0.0

# データバージョン管理
dvc>=3.0.0

# ハイパーパラメータチューニング
optuna>=3.0.0

# データ処理
albumentations>=1.3.0
Pillow>=10.0.0
opencv-python>=4.8.0

# テスト
pytest>=7.0.0
pytest-cov>=4.0.0

# 可視化
matplotlib>=3.7.0

# Webフレームワーク
django>=4.2.0
pillow>=10.0.0  # Django ImageField用（既に含まれているが明記）

# その他
pyyaml>=6.0
numpy>=1.24.0,<2.0.0
optuna-integration[pytorch_lightning]>=0.9.0
